build(deps): migrate docs to fumadocs-openapi v11 and bump docs-npm group
fumadocs-openapi v11 removes the RSC createAPIPage(openapi, opts) API (only the
deprecated createClientAPIPage alias survives), so the grouped Dependabot bump
(#454) failed type-checking on lib/openapi.ts. Migrate to the v11 client/server
split: lib/openapi-page.tsx is the "use client" component built with
createOpenAPIPage (shiki factory now defaults internally), and lib/openapi.tsx
is an async server component that loads the bundled schema and forwards
payload.bundled + operations to it.

This lands the rest of the #454 group too (fumadocs-core 16.10, fumadocs-mdx
15.0.12, next 16.2.9, @types/*); fumadocs-ui/lucide-react move within range in
the lockfile.

Verified: tsc and next build pass, and a headless-Chrome render of the
url_shortener API viewer shows all five operations with the interactive
playground and zero console/hydration errors.
