+++ docs/package.json
@@ -18,13 +18,13 @@
     "@orama/orama": "^3.1.8",
     "@shikijs/transformers": "^4.2.0",
     "beautiful-mermaid": "^1.1.3",
-    "fumadocs-core": "^16.9.3",
-    "fumadocs-mdx": "^15.0.10",
-    "fumadocs-openapi": "^10.10.3",
+    "fumadocs-core": "^16.10.0",
+    "fumadocs-mdx": "^15.0.12",
+    "fumadocs-openapi": "^11.0.0",
     "fumadocs-ui": "^16.9.3",
     "katex": "^0.17.0",
     "lucide-react": "^1.16.0",
-    "next": "^16.2.7",
+    "next": "^16.2.9",
     "railroad-diagrams": "^1.0.0",
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
@@ -37,9 +37,9 @@
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.3.0",
-    "@types/mdx": "^2.0.13",
-    "@types/node": "25.9.1",
-    "@types/react": "^19.2.16",
+    "@types/mdx": "^2.0.14",
+    "@types/node": "25.9.3",
+    "@types/react": "^19.2.17",
     "@types/react-dom": "^19.1.2",
     "postcss": "^8.5.15",
     "tailwindcss": "^4.2.4",
